Consolidate LocationDisplay and Test Component
