+++ src/applications/vass/containers/withFlowGuard.unit.spec.jsx
@@ -1,28 +1,16 @@
 import React from 'react';
 import { expect } from 'chai';
 import { waitFor } from '@testing-library/react';
-import { Routes, Route, useLocation } from 'react-router-dom-v5-compat';
+import { Routes, Route } from 'react-router-dom-v5-compat';
 import { renderWithStoreAndRouterV6 } from '~/platform/testing/unit/react-testing-library-helpers';
 
 import withFlowGuard from './withFlowGuard';
 import { FLOW_TYPES, URLS } from '../utils/constants';
-import { getDefaultRenderOptions } from '../utils/test-utils';
-
-// Helper component to display current location for testing navigation
-const LocationDisplay = () => {
-  const location = useLocation();
-  return (
-    <div data-testid="location-display">
-      {location.pathname}
-      {location.search}
-    </div>
-  );
-};
-
-// Simple test component to wrap
-const TestComponent = () => (
-  <div data-testid="test-component">Test Content</div>
-);
+import {
+  getDefaultRenderOptions,
+  LocationDisplay,
+  TestComponent,
+} from '../utils/test-utils';
 
 describe('VASS Containers: withFlowGuard', () => {
   describe('with FLOW_TYPES.ANY (default)', () => {
