+++ src/applications/vass/containers/withFormData.unit.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import { waitFor } from '@testing-library/react';
-import { Routes, Route, useLocation } from 'react-router-dom-v5-compat';
+import { Routes, Route } from 'react-router-dom-v5-compat';
 import { renderWithStoreAndRouterV6 } from '~/platform/testing/unit/react-testing-library-helpers';
 
 import withFormData from './withFormData';
@@ -11,23 +11,9 @@ import * as formSlice from '../redux/slices/formSlice';
 import {
   getDefaultRenderOptions,
   getHydratedFormRenderOptions,
+  LocationDisplay,
+  TestComponent,
 } from '../utils/test-utils';
-
-// Helper component to display current location for testing navigation
-const LocationDisplay = () => {
-  const location = useLocation();
-  return (
-    <div data-testid="location-display">
-      {location.pathname}
-      {location.search}
-    </div>
-  );
-};
-
-// Simple test component to wrap
-const TestComponent = () => (
-  <div data-testid="test-component">Test Content</div>
-);
 
 describe('VASS Containers: withFormData', () => {
   let loadFormDataFromStorageStub;
